TODO: TrackScheduler

+++ src/main/java/org/musicbox/listeners/CommandListener.java
@@ -1,15 +1,10 @@
 package org.musicbox.listeners;
 
-import java.util.List;
 
 import org.musicbox.config.DefaultConfig;
 import org.musicbox.core.Permissions;
 import org.musicbox.core.managers.CommandManager;
 import org.musicbox.core.models.Listener;
-import org.musicbox.core.utils.Constants;
-import org.musicbox.core.utils.Messages;
-import org.musicbox.core.utils.Placeholder;
-import org.musicbox.core.utils.PlaceholderBuilder;
 import org.musicbox.core.utils.Utils;
 
 import net.dv8tion.jda.api.entities.ChannelType;
@@ -23,17 +18,13 @@ public class CommandListener extends Listener {
       if (!event.isFromType(ChannelType.TEXT) || event.getAuthor().isBot())
          return;
 
-      List<Placeholder> placeholders = PlaceholderBuilder.createBy(event, true)
-            .add(Constants.KEY_MISSING_PERMISSIONS, Utils.toString(Permissions.WRITING_PERMISSIONS)).build();
-
       /* check if the bot can send embeds in the text channel */
-      if (!Permissions.canWrite(event.getTextChannel(), event.getMember())) {
-         Messages.translatedMessage(event, Messages.COMMAND_MISSING_PERMISSION, placeholders);
+      if (!Permissions.canWrite(event.getTextChannel(), event.getGuild().getSelfMember())) {
          return;
       }
 
       /* kernel's id */
-      if (event.getMember().getUser().getIdLong() != 339978701297156098L) {
+      if (DefaultConfig.DEBUG_MODE && event.getMember().getUser().getIdLong() != 339978701297156098L) {
          event.getTextChannel().sendMessage("Você não tem permissão de usar o bot em modo de desenvolvimento.")
                .queue(Utils.deleteAfter(20L));
          return;
